refact(decman-lib): move the record toolkit
One shared module now holds both record-reading shapes: the strict
Result readers and the lenient Option readers, plus the Set Party,
GenMap, and RelTime extractors. Sources: decman record.rs, the
action_serializer extract helpers, and the queries field helpers.
