+++ crates/decman-lib/src/framework/mod.rs
@@ -1,6 +1,7 @@
 //! Integrator-facing, template-agnostic pieces.
 
 pub mod encode;
+pub mod record;
 
 mod template_id;
 pub use template_id::TemplateId;
